requirements.txt
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.9
Cython==0.29.25
dataclasses==0.8
decorator==4.4.2
distlib==0.3.4
filelock==3.4.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
isodate==0.6.0
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
networkx==2.5.1
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.19.1
pylzma==0.5.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
rdflib==5.0.0
requests==2.26.0
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
tensorboardX==2.4.1
threadpoolctl==3.0.0
torch==1.7.1+cu101
torch-geometric==2.0.2
torch-scatter==2.0.7
torch-sparse==0.6.12
torchaudio==0.7.2
torchvision==0.8.2+cu101
tqdm==4.62.3
typing-extensions==4.0.1
yacs==0.1.8
zipp==3.6.0
